refactor(es): Rename the dead-letter connector to jaeger_storage_writer
The connector's primary job is the synchronous storage write; dead-lettering is
an exceptional tap on its output. Naming it jaeger_storage_dead_letter made the
main pipeline read backwards — `exporters: [jaeger_storage_dead_letter]` naively
scans as "export to a dead letter" when it actually writes to storage. Rename the
type (and package) to jaeger_storage_writer so the exporter slot reads as "write
to storage"; the dead-letter role is carried by the output pipeline it feeds
(traces/dead_letter), not the component name.

Also rebased onto main to fold in #9113: classify() now returns the 409-as-
idempotent-conflict tally alongside the []RejectedItem terminal set the connector
needs, so a retried-and-already-durable span is never mis-routed to dead-letter.
