+++ cmd/jaeger/internal/connectors/storagewriterconnector/package_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package storagedeadletterconnector
+package storagewriterconnector
 
 import (
 	"testing"
